Better keyboard accessibility of nav modules
I wrote a couple of dozen lines of JavaScript to handle when you should
or should not be able to tab into a menu, but then I realized that css
"visibility: hidden" does everything for you. Anything in a hidden
element is not tabbable or focusable. So that was really all I needed.
Go go learning basic browser functionality.

+ add .invisible standard class to set elements to visibility:hidden
+ Non-active panels in the Whirligig carousel are .invisible. They are
all set to visible when the carousel is inactive.
+ The main navigation module sets the main menu to hidden when it
properly should be (small and closed) and visible all other times.
+ Changed some transitions on Bartender

Related to issue #2 and issue #3