Various changes to make elements fit
+ Remove automatic word breaking on pre/code blocks
+ Add utility classes for overflow and word breaking
+ Remove white-space pre-wrap from pre,code,tt elements
+ Remove overflow:hidden from .off-canvas-wrap. I thought this was
necessary but it doesn't appear to really be. Its presence could cause
some real damage in trying to read big tables
+ New JS module: "Compactor". It looks for pre/tables that are too
large and wraps them in an element that allows overflow. It also looks
for errant code/anchors that are too wide and sets them to break lines.
This runs on load and window resize

+ (unrelated) Headers in .copy are now normal weight instead of bold.
They stand out enough with the borders

Close #18